# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Freelancers",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://api-freelancer.otterscompany.com"
        }
    ],
    "paths": {
...
        "/api/v1/proposals": {
            "get": {
                "tags": [
                    "Proposals"
                ],
                "summary": "Get proposals",
                "description": "Returns all proposals",
                "operationId": "indexProposal",
                "responses": {
                    "200": {
                        "description": "Show all proposals.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/proposals/{id}": {
            "get": {
                "tags": [
                    "Proposals"
                ],
                "summary": "Get proposal",
                "description": "Returns a single proposal",
                "operationId": "showProposal",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Search for ID proposal",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show proposal.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    },
                    "404": {
                        "description": "No results found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Proposals"
                ],
                "summary": "Update proposal",
                "description": "Update a single proposal",
                "operationId": "updateProposal",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Search for ID proposal",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start",
                        "in": "query",
                        "description": "Date start of project, format Y-m-d",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end",
                        "in": "query",
                        "description": "Date end of project, format Y-m-d",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "amount",
                        "in": "query",
                        "description": "Amount of proposal",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "Content of proposal",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update proposal.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    },
                    "403": {
                        "description": "Forbidden."
                    },
                    "422": {
                        "description": "Data not valid."
                    },
                    "500": {
                        "description": "An error occurred during the process."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Proposals"
                ],
                "summary": "Register proposal",
                "description": "Create a new proposal",
                "operationId": "storeProposal",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Chat ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start",
                        "in": "query",
                        "description": "Date start of project, format Y-m-d",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end",
                        "in": "query",
                        "description": "Date end of project, format Y-m-d",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "amount",
                        "in": "query",
                        "description": "Amount of proposal",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "Content of proposal",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "receiver_id",
                        "in": "query",
                        "description": "User to whom the proposal is sent",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Registered proposal.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    },
                    "422": {
                        "description": "Data not valid."
                    },
                    "500": {
                        "description": "An error occurred during the process."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/proposals/{id}/cancel": {
            "put": {
                "tags": [
                    "Proposals"
                ],
                "summary": "Cancel proposal",
                "description": "Cancel a single proposal",
                "operationId": "cancelProposal",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Search for ID proposal",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cancel proposal.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    },
                    "403": {
                        "description": "Forbidden."
                    },
                    "404": {
                        "description": "No results found."
                    },
                    "500": {
                        "description": "An error occurred during the process."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/proposals/{id}/accept": {
            "put": {
                "tags": [
                    "Proposals"
                ],
                "summary": "Accept proposal",
                "description": "Accept a single proposal",
                "operationId": "acceptProposal",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Search for ID proposal",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Accept proposal.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    },
                    "403": {
                        "description": "Forbidden."
                    },
                    "404": {
                        "description": "No results found."
                    },
                    "500": {
                        "description": "An error occurred during the process."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}